requirements.txt
amqp==2.2.1
Babel==2.4.0
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
coreapi==2.3.1
coreapi-cli==1.0.6
coreschema==0.0.4
Django==1.11.4
django-filter==1.0.4
django-rest-swagger==2.1.2
djangorestframework==3.6.3
docker==2.5.0
docker-pycreds==0.2.1
flower==0.9.2
futures==3.1.1
idna==2.6
ipaddress==1.0.18
itypes==1.1.0
Jinja2==2.9.6
kombu==4.1.0
Markdown==2.6.8
MarkupSafe==1.0
openapi-codec==1.3.2
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.10.0
slumber==0.7.1
tornado==4.5.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
websocket-client==0.44.0
